build-with-ai-creating-ai-agents-with-gpt-5/requirements.txt
python-dotenv
fastapi
requests
uvicorn
google-genai
